moved all my work into the dev branch, added it to all the routes and got the navbar buttons working.
