+++ package.json
@@ -5,15 +5,12 @@
   "dependencies": {
     "@material-ui/core": "^3.9.0",
     "@material-ui/icons": "^3.0.2",
-
     "axios": "^0.18.0",
     "dot-env": "0.0.1",
     "express": "^4.16.4",
     "googleapis": "^36.0.0",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.3",
     "react-redux": "^6.0.0",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.3",
