+++ src/Components/App.js
@@ -6,6 +6,9 @@ import NewsFeedPage from "./pages/NewsFeedPage";
 import LoginPage from "./pages/LoginPage";
 import PrivateRoute from "./PrivateRoute";
 import NewChallengePage from "./pages/NewChallengePage";
+import FeedPage from "./pages/FeedPage";
+import ProfilePage from "./pages/ProfilePage";
+import { Switch } from "react-router-dom";
 
 class App extends Component {
   
@@ -13,13 +16,17 @@ class App extends Component {
     return (
       <div>
         <BrowserRouter>
-          <div>
-            <Route exact path = "/portal" component = {MainPage} />
-            <Route exact path = "/login" component = {LoginPage} />
-            <Route exact path = "/register" component = {RegisterPage} />
-            <PrivateRoute exact path = "/" component = {NewsFeedPage} />
-            <PrivateRoute exact path = "/newchallenge" component = {NewChallengePage} />
-          </div>
+            <div>
+              <Switch>
+                <Route exact path = "/portal" component = {MainPage} />
+                <Route exact path = "/login" component = {LoginPage} />
+                <Route exact path = "/register" component = {RegisterPage} />
+                <Route exact path = "/feed" component = {FeedPage} />
+                <Route exact path = "/profile" component = {ProfilePage} />
+                <PrivateRoute exact path = "/" component = {NewsFeedPage} />
+                <PrivateRoute exact path = "/newchallenge" component = {NewChallengePage} />
+              </Switch>
+            </div>
         </BrowserRouter>
       </div>
     );
