+++ src/Components/pages/NewsFeedPage.js
@@ -1,6 +1,8 @@
 import React, { Component } from "react";
 import { removeAuthToken } from "./../../actions";
 import { connect } from "react-redux"; 
+import NavBar from '../components/NavBar';
+import ChallengeCard from '../components/ChallengeCard';
 
 class NewsFeedPage extends Component  {
 
@@ -8,16 +10,35 @@ class NewsFeedPage extends Component  {
         const { removeAuthToken } = this.props;
 
         return (
+            
             <div>
-                <h1> News Feed</h1>
-                <button onClick={() => {
-                    removeAuthToken();
-                }}> Logout </button>
+                <div>
+                    <NavBar />
+                    <div class="challenge_card">
+                    <ChallengeCard />
+                    </div>
+                    <div class="challenge_card">
+                    <ChallengeCard />
+                    <div class="challenge_card">
+                    <ChallengeCard />
+                    </div>
+                </div>
             </div>
+            </div>
+            
         );
     }
 }
 
-export default connect(null, {
-    removeAuthToken
-})(NewsFeedPage);
+export default NewsFeedPage;
+
+// export default connect(null, {
+//     removeAuthToken
+// })(NewsFeedPage);
+
+// </div>
+// <h1> News Feed</h1>
+// <button onClick={() => {
+//     removeAuthToken();
+// }}> Logout </button>
+// </div>
