+++ src/components/components/ChallengeCard.js
@@ -16,7 +16,7 @@ import FavoriteIcon from '@material-ui/icons/Favorite';
 import ShareIcon from '@material-ui/icons/Share';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import MoreVertIcon from '@material-ui/icons/MoreVert';
-import Snowboard from './../images/snowboarding.jpg';
+import Snowboard from './images/snowboarding.jpg';
 
 const styles = theme => ({
   card: {
