+++ src/components/components/ProfileInfoCard.js
@@ -7,7 +7,7 @@ import CardContent from '@material-ui/core/CardContent';
 import Typography from '@material-ui/core/Typography';
 import CardMedia from '@material-ui/core/CardMedia';
 import Grid from '@material-ui/core/Grid';
-import photo from './../images/profile_image.jpg';
+import photo from './images/profile_image.jpg';
