+++ src/components/pages/ProfilePage.js
@@ -1,23 +1,23 @@
 import React, { Component } from 'react';
-import NavBar from './components/NavBar';
-import ChallengeCard from './components/ChallengeCard';
+import './../App.css';
+import NavBar from '../components/NavBar';
+import ProfileInfoCard from '../components/ProfileInfoCard';
+import ChallengeCard from '../components/ChallengeCard';
+
 
 class Profile extends Component {
   render() {
     return (
       <div>
         <NavBar />
-        <div class="challenge_card">
-         <ChallengeCard />
+        <div class="profile_card">
+          <ProfileInfoCard />
         </div>
         <div class="challenge_card">
          <ChallengeCard />
-         <div class="challenge_card">
-         <ChallengeCard />
-        </div>
         </div>
       </div>
-     
+      
     );
   }
 }
